+++ tests/Runtime/AgentFactoryTest.php
@@ -8,6 +8,7 @@ use BEAR\Resource\FactoryInterface;
 use BEAR\Resource\Module\ResourceModule;
 use BEAR\Resource\ResourceInterface;
 use BEAR\ToolUse\Dispatch\Dispatcher;
+use BEAR\ToolUse\Dispatch\NullToolCallObserver;
 use BEAR\ToolUse\Dispatch\ToolRegistry;
 use BEAR\ToolUse\Fake\FakeConfirmationHandler;
 use BEAR\ToolUse\Fake\FakeLlmClient;
@@ -35,7 +36,7 @@ final class AgentFactoryTest extends TestCase
         $registry = new ToolRegistry();
         $converter = new SchemaConverter(DocBlockFactory::createInstance());
         $collector = new ToolCollector($converter, $registry, $resourceFactory);
-        $dispatcher = new Dispatcher($resource, $registry);
+        $dispatcher = new Dispatcher($resource, $registry, new NullToolCallObserver());
 
         $this->factory = new AgentFactory($this->llmClient, $dispatcher, $collector, $registry);
     }
@@ -108,7 +109,7 @@ final class AgentFactoryTest extends TestCase
         $registry   = new ToolRegistry();
         $converter  = new SchemaConverter(DocBlockFactory::createInstance());
         $collector  = new ToolCollector($converter, $registry, $resourceFactory);
-        $dispatcher = new Dispatcher($resource, $registry);
+        $dispatcher = new Dispatcher($resource, $registry, new NullToolCallObserver());
 
         $factory = new AgentFactory($this->llmClient, $dispatcher, $collector, $registry, null, $streamingClient);
         $factory->addResources(['app://self/article']);
@@ -136,7 +137,7 @@ final class AgentFactoryTest extends TestCase
         $registry  = new ToolRegistry();
         $converter = new SchemaConverter(DocBlockFactory::createInstance());
         $collector = new ToolCollector($converter, $registry, $resourceFactory);
-        $dispatcher = new Dispatcher($resource, $registry);
+        $dispatcher = new Dispatcher($resource, $registry, new NullToolCallObserver());
 
         $factory = new AgentFactory($this->llmClient, $dispatcher, $collector, $registry, $confirmationHandler);
         $factory->addResources(['app://self/article']);
@@ -163,7 +164,7 @@ final class AgentFactoryTest extends TestCase
         $registry   = new ToolRegistry();
         $converter  = new SchemaConverter(DocBlockFactory::createInstance());
         $collector  = new ToolCollector($converter, $registry, $resourceFactory);
-        $dispatcher = new Dispatcher($resource, $registry);
+        $dispatcher = new Dispatcher($resource, $registry, new NullToolCallObserver());
 
         $factory = new AgentFactory(
             $this->llmClient,